Use libtool to handle libraries Fixed to build with GCC 3
git-svn-id: file:///home/xapian-svn/xapian/trunk@4942 6b2f1b92-a4f3-0310-943f-de3cbe41748b

+++ xapian-applications/cvssearch/map/range/range.cpp
@@ -181,7 +181,7 @@ range::operator-(const range & r)
             (*this) = range(intersect.end(), _end-1);
         }
     }
-    return range(0,-1);
+    return range(0,(unsigned int)-1);
 }
     
 bool
